add eclipse project settings
The .project and .settings files were left out of previous commits but
they are needed for project meta-data. I have started tracking the files
and updated the gitignore file not to ignore them. It should now be
easier for people wanting to join the project to clone and import the
project.